Add fixes according to .proto file

+++ internal/adapters/api/interfaces.go
@@ -6,10 +6,10 @@ import (
 )
 
 type Service interface {
-	GetPostById(ctx context.Context, id int64) (*domain.Post, error)
-	GetPostsByUserId(ctx context.Context, id int64) ([]domain.Post, error)
+	GetPostById(ctx context.Context, id int32) (*domain.Post, error)
+	GetPostsByUserId(ctx context.Context, id int32) ([]domain.Post, error)
 	GetAllPosts(ctx context.Context) ([]domain.Post, error)
 	Save(ctx context.Context, user domain.SavePostDTO) (*domain.Post, error)
 	Update(ctx context.Context, user domain.UpdatePostDTO) error
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id int32) error
 }
