+++ internal/domain/dto.go
@@ -1,13 +1,13 @@
 package domain
 
 type SavePostDTO struct {
-	UserId int64  `json:"user_id"`
+	UserId int32  `json:"user_id"`
 	Title  string `json:"title"`
 	Body   string `json:"body"`
 }
 
 type UpdatePostDTO struct {
-	Id    int64  `json:"id"`
+	Id    int32  `json:"id"`
 	Title string `json:"title"`
 	Body  string `json:"body"`
 }
