+++ internal/domain/model.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Post struct {
-	Id     int64  `json:"id"`
-	UserId int64  `json:"user_id omitempty"`
+	Id     int32  `json:"id"`
+	UserId int32  `json:"user_id omitempty"`
 	Title  string `json:"title omitempty"`
 	Body   string `json:"body omitempty"`
 }
